Extract transaction creation into factory class
